@@ -3,28 +3,26 @@
   "version": "0.0.0",
   "author": "Danes je nov dan",
   "license": "Unlicense",
+  "type": "module",
   "scripts": {
     "dev": "nodemon server/index.js",
     "lint": "eslint .",
     "start": "node server/index.js"
   },
   "dependencies": {
-    "body-parser": "^1.15.2",
-    "chalk": "^2.4.1",
-    "cron": "^1.5.1",
-    "cross-env": "^5.2.0",
-    "ejs": "^3.1.7",
-    "express": "^4.14.0",
-    "fs-extra": "^7.0.0",
-    "lodash": "^4.17.21",
-    "node-fetch": "^2.6.7",
-    "serve-static": "^1.13.2",
-    "url-pattern": "^1.0.1"
+    "ejs": "^3.1.8",
+    "express": "^4.18.2",
+    "fs-extra": "^10.1.0",
+    "lodash-es": "^4.17.21",
+    "node-fetch": "^3.2.10"
   },
   "devDependencies": {
-    "eslint": "4.19.1",
-    "eslint-config-airbnb-base": "13.0.0",
-    "eslint-plugin-import": "2.12.0",
-    "nodemon": "^2.0.15"
+    "eslint": "^8.25.0",
+    "eslint-config-airbnb-base": "15.0.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "nodemon": "^2.0.20",
+    "prettier": "^2.7.1"
   }
 }
