@@ -1,14 +1,12 @@
-const chalk = require('chalk');
-const server = require('./server');
+import { setupExpress } from './server.js';
 
-Promise.resolve()
-  .then(server.init)
+setupExpress()
   .then(() => {
     // eslint-disable-next-line no-console
-    console.log(chalk.green('All is well!'));
+    console.log('All is well!');
   })
   .catch((error) => {
     // eslint-disable-next-line no-console
-    console.error(chalk.red('Failed to start:'), error);
+    console.error('Failed to start:', error);
     process.exit(1);
   });
