@@ -1,13 +1,13 @@
-const config = {
+export default {
   port: 3066,
   serverTimeout: 120000,
   urls: {
-    // cdn: 'http://localhost:8080',
-    // cards: 'http://localhost:7004',
-    // parladata: 'http://localhost:8000/v3',
-    cdn: 'https://parlassets-ljubljana.lb.djnd.si',
-    cards: 'https://parlacards.lb.djnd.si',
-    parladata: 'https://parladata.lb.djnd.si/v3',
+    cdn: 'http://localhost:8080',
+    cards: 'http://localhost:7004',
+    parladata: 'http://localhost:8000/v3',
+    // cdn: 'https://parlassets-ljubljana.lb.djnd.si',
+    // cards: 'https://parlacards.lb.djnd.si',
+    // parladata: 'https://parladata.lb.djnd.si/v3',
     metaImages: 'https://meta-image-generator.lb.djnd.si/parlameter',
   },
   locale: 'sl-obcina-ljubljana',
@@ -17,9 +17,3 @@ const config = {
   defaultCardDate: process.env.PARLASITE_DEFAULT_CARD_DATE,
   newsletterSegmentId: process.env.PARLASITE_NEWSLETTER_SEGMENT_ID,
 };
-
-if (!config.leaderId || !config.rootOrgId || !config.mandateId) {
-  throw new Error('Required config values are not defined!');
-}
-
-module.exports = config;
